feat: address and hash concatination refactor
